Rename Red wines from Asia.csv to Red wines dataset (from Asia).csv
The dataset contains the physicochemical properties of Asian wines, scraped over approximately 30 hours. The purpose of this dataset is to identify the representative characteristics of Silver, Gold, and Grand Gold medal-winning wines. This information could be useful for winemakers, wine enthusiasts, and researchers interested in understanding the attributes of high-quality Asian wines. The variables in the dataset include physical and chemical measurements such as pH, alcohol content, sugar levels, and others that contribute to the overall quality and taste of the wine. With this data, it is possible to perform statistical analyses and identify patterns and correlations between the variables and the medal categories. This dataset represents a valuable resource for anyone seeking to gain insight into the properties of high-quality Asian wines.